# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "EventTemplateCreate": {
        "description": "Schema for creating an event template.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "template_json": {
            "title": "Template Json",
            "type": "string"
          }
        },
        "required": [
          "name",
          "template_json"
        ],
        "title": "EventTemplateCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PollCreate": {
        "description": "Create a poll.",
        "properties": {
          "allow_multiple": {
            "default": false,
            "title": "Allow Multiple",
            "type": "boolean"
          },
          "close_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "is_anonymous": {
            "default": false,
            "title": "Is Anonymous",
            "type": "boolean"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/PollOptionCreate"
            },
            "title": "Options",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "options"
        ],
        "title": "PollCreate",
        "type": "object"
      },
...
      "PollTemplateCreate": {
        "description": "Create a custom poll template.",
        "properties": {
          "allow_multiple": {
            "default": false,
            "title": "Allow Multiple",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          }
        },
        "required": [
          "name",
          "options"
        ],
        "title": "PollTemplateCreate",
        "type": "object"
      },
...
      "UpdateFamilySettingsRequest": {
        "description": "Update family settings request.",
        "properties": {
          "theme_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme Color"
          }
        },
        "title": "UpdateFamilySettingsRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Family event coordination platform",
    "title": "FamilyCircle",
    "version": "3.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/event-templates": {
      "get": {
        "description": "List event templates for the current family.",
        "operationId": "list_templates_api_event_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Templates",
        "tags": [
          "event-templates"
        ]
      },
      "post": {
        "description": "Create a new event template. Requires family admin.",
        "operationId": "create_template_api_event_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Template",
        "tags": [
          "event-templates"
        ]
      }
    },
    "/api/event-templates/{template_id}": {
      "delete": {
        "description": "Delete an event template. Requires family admin.",
        "operationId": "delete_template_api_event_templates__template_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Template",
        "tags": [
          "event-templates"
        ]
      }
    },
    "/api/events": {
      "get": {
        "description": "List all events for current family.",
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Create a new event in current family.",
        "operationId": "create_event_api_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/events/{event_id}": {
      "delete": {
        "description": "Delete an event (creator or admin only).",
        "operationId": "delete_event_api_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Event",
        "tags": [
          "events"
        ]
      },
      "get": {
        "description": "Get a specific event.",
        "operationId": "get_event_api_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Event",
        "tags": [
          "events"
        ]
      },
      "put": {
        "description": "Update an event (creator or admin only).",
        "operationId": "update_event_api_events__event_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/gift-exchange/{event_id}": {
      "get": {
        "description": "Get Gift Exchange info for an event.",
        "operationId": "get_gift_exchange_api_gift_exchange__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Gift Exchange",
        "tags": [
          "gift-exchange"
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "description": "Get all notification settings (super admin only).\n\nSensitive values (tokens, webhooks) are masked.",
        "operationId": "get_notification_settings_api_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Notification Settings",
        "tags": [
          "notifications"
        ]
      },
      "put": {
        "description": "Update notification settings (super admin only).\n\nMasked values (********) are ignored to avoid overwriting secrets.",
        "operationId": "update_notification_settings_api_notifications_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Notification Settings",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/poll-templates": {
      "get": {
        "description": "List available poll templates (built-in + family custom).",
        "operationId": "list_templates_api_poll_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Templates",
        "tags": [
          "poll-templates"
        ]
      },
      "post": {
        "description": "Create a custom poll template for the current family.",
        "operationId": "create_template_api_poll_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Template",
        "tags": [
          "poll-templates"
        ]
      }
    },
    "/api/poll-templates/{template_id}": {
      "delete": {
        "description": "Delete a custom poll template. Built-in templates cannot be deleted.",
        "operationId": "delete_template_api_poll_templates__template_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "template_id",
            "required": true,
            "schema": {
              "title": "Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Template",
        "tags": [
          "poll-templates"
        ]
      }
    },
    "/api/polls": {
      "get": {
        "description": "List polls for the current family, optionally filtered by event.",
        "operationId": "list_polls_api_polls_get",
        "parameters": [
          {
            "in": "query",
            "name": "event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Polls",
        "tags": [
          "polls"
        ]
      },
      "post": {
        "description": "Create a new poll.",
        "operationId": "create_poll_api_polls_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Poll",
        "tags": [
          "polls"
        ]
      }
    },
    "/api/polls/{poll_id}": {
      "delete": {
        "description": "Delete a poll. Requires creator or family admin.",
        "operationId": "delete_poll_api_polls__poll_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "poll_id",
            "required": true,
            "schema": {
              "title": "Poll Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Poll",
        "tags": [
          "polls"
        ]
      },
      "get": {
        "description": "Get a single poll with full details.",
        "operationId": "get_poll_api_polls__poll_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "poll_id",
            "required": true,
            "schema": {
              "title": "Poll Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Poll",
        "tags": [
          "polls"
        ]
      }
    },
...
    "/api/potluck/{event_id}": {
      "get": {
        "description": "Get potluck info for an event.",
        "operationId": "get_potluck_api_potluck__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Potluck",
        "tags": [
          "potluck"
        ]
      }
    },
...
    "/api/profile": {
      "get": {
        "description": "Get current user's profile.",
        "operationId": "get_profile_api_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "profile"
        ]
      },
      "put": {
        "description": "Update current user's profile.",
        "operationId": "update_profile_api_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "profile"
        ]
      }
    },
...
    "/api/settings": {
      "get": {
        "description": "Get application settings.\n\n- Public: app_name\n- Authenticated: family-specific settings\n- Super admin: SMTP and other sensitive global settings",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "description": "Update settings (backwards compatibility, redirects to global).",
        "operationId": "update_settings_compat_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGlobalSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Settings Compat",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/family": {
      "put": {
        "description": "Update family-specific settings (family admin only).",
        "operationId": "update_family_settings_api_settings_family_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFamilySettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Family Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/wishlist": {
      "get": {
        "description": "Get current user's wishlist.",
        "operationId": "get_wishlist_api_wishlist_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Wishlist",
        "tags": [
          "wishlist"
        ]
      },
      "post": {
        "description": "Add an item to wishlist.",
        "operationId": "add_wishlist_item_api_wishlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishlistItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Wishlist Item",
        "tags": [
          "wishlist"
        ]
      }
    },
...
  }
}